% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearish.candle.R
\name{bearish.candle}
\alias{bearish.candle}
\title{Determine bearish candle using a OHLC price series}
\usage{
bearish.candle(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candel detected
}
\description{
Determine bearish candle using a OHLC price series
}
