% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearish.engulf.R
\name{bearish.engulf}
\alias{bearish.engulf}
\title{Determine bearish engulfing pattern using a OHLC price series}
\usage{
bearish.engulf(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bearish engulfing pattern detected
}
\description{
Determine bearish engulfing pattern using a OHLC price series
}
