% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearish.harami.R
\name{bearish.harami}
\alias{bearish.harami}
\title{Determine bearish harami pattern using a OHLC price series}
\usage{
bearish.harami(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bearish haramipattern detected
}
\description{
Determine bearish harami pattern using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
