% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_harami.R
\name{bearish_harami}
\alias{bearish_harami}
\title{Determine bearish harami pattern using a OHLC price series}
\usage{
bearish_harami(ohlc)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bearish haramipattern detected
}
\description{
Determine bearish harami pattern using a OHLC price series
}
