% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullish.candle.R
\name{bullish.candle}
\alias{bullish.candle}
\title{Determine bullish candle using a OHLC price series}
\usage{
bullish.candle(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candel detected
}
\description{
Determine bullish candle using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
