% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullish.engulf.R
\name{bullish.engulf}
\alias{bullish.engulf}
\title{Determine bullish engulfing pattern using a OHLC price series}
\usage{
bullish.engulf(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if hammer pattern detected
}
\description{
Determine bullish engulfing pattern using a OHLC price series
}
