% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullish.harami.R
\name{bullish.harami}
\alias{bullish.harami}
\title{Determine bullish harami pattern using a OHLC price series}
\usage{
bullish.harami(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish harami  pattern detected
}
\description{
Determine bullish harami pattern using a OHLC price series
}
