% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_engulfing.R
\name{bullish_engulfing}
\alias{bullish_engulfing}
\title{Determine bullish engulfing pattern using a OHLC price series}
\usage{
bullish_engulfing(ohlc)
}
\arguments{
\item{ohlc}{A data.frame containing Open, High, Low, Close prices}
}
\value{
Returns a data.frame-object with identified patterns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other engulfing patterns: 
\code{\link{bearish_engulfing}()}

Other 2-day patterns: 
\code{\link{bearish_engulfing}()},
\code{\link{bearish_harami}()},
\code{\link{bullish_harami}()}
}
\author{
Chiu-Yu Ko
}
\concept{2-day patterns}
\concept{engulfing patterns}
