% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_harami.R
\name{bullish_harami}
\alias{bullish_harami}
\title{Determine bullish harami pattern using a OHLC price series}
\usage{
bullish_harami(ohlc)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish harami  pattern detected
}
\description{
Determine bullish harami pattern using a OHLC price series
}
