% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns_fn.R
\name{candle_pattern}
\alias{candle_pattern}
\title{Find Candlestick Patterns}
\usage{
candle_pattern(pattern, ohlc)
}
\arguments{
\item{pattern}{A \link{list} of candlestick patterns}

\item{ohlc}{A two-dimensional tabular data.frame with Open, High, Low and Close prices.}
}
\value{
An object of same \link{class}, with candlestick pattern indicators
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link[=candle_pattern]{candle_pattern()}}-function returns
}
\examples{
# script: Candlestick Patters
# date: 2024-02-09
# author: Serkan Korkmaz, serkor1@duck.com
# objective: Create an example of the
# usage of the function candle_pattern
# script start;

# garbage collection
rm(list = ls()); invisible(gc());

# library
library(candlestickpatterns)

# 1) add pattern indicators
# to MSFT
MSFT <- candle_pattern(
  ohlc    = MSFT,
  pattern = list(
    # 1) Bullish patterns
    bullish_engulfing(),
    bullish_harami(),

    # 2) Bearish patterns
    bearish_engulfing(),
    bearish_harami()

  )
)

# 2) head data
head(
  MSFT
)

# script end;
}
\author{
Serkan Korkmaz
}
