% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleBodyBottom.R
\name{CandleBodyBottom}
\alias{CandleBodyBottom}
\title{Determine candle body bottom using a OHLC price series}
\usage{
CandleBodyBottom(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
bottom of candle body
}
\description{
Determine candle body bottom using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
