% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleBodyCenter.R
\name{CandleBodyCenter}
\alias{CandleBodyCenter}
\title{Determine candle body center using a OHLC price series}
\usage{
CandleBodyCenter(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
center of candle body
}
\description{
Determine candle body center using a OHLC price series
}
