% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleBodyLength.R
\name{CandleBodyLength}
\alias{CandleBodyLength}
\title{Determine candle body length using a OHLC price series}
\usage{
CandleBodyLength(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
length of candle body
}
\description{
Determine candle body length using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
