% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleBodyTop.R
\name{CandleBodyTop}
\alias{CandleBodyTop}
\title{Determine candle body top using a OHLC price series}
\usage{
CandleBodyTop(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
top of candle body
}
\description{
Determine candle body top using a OHLC price series
}
