% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandleStickLength.R
\name{CandleStickLength}
\alias{CandleStickLength}
\title{Determine candle stick length using a OHLC price series}
\usage{
CandleStickLength(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if Dragon Doji pattern detected
}
\description{
Determine candle stick length using a OHLC price series
}
