% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder-package.R
\docType{package}
\name{candlestickpatterns-package}
\alias{candlestickpatterns}
\alias{candlestickpatterns-package}
\title{candlestickpatterns: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chiu-yu Ko \email{chiuyuko@cuhk.edu.hk} (\href{https://orcid.org/0000-0001-8590-6159}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Serkan Korkmaz \email{serkor1@duck.com} (\href{https://orcid.org/0000-0002-5052-0982}{ORCID})
}

}
\keyword{internal}
