% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark.cloud.cover.R
\name{dark.cloud.cover}
\alias{dark.cloud.cover}
\title{Determine dark cloud cover pattern using a OHLC price series}
\usage{
dark.cloud.cover(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if dark cloud cover pattern detected
}
\description{
Determine dark cloud cover pattern using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
