% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_cloud_cover.R
\name{dark_cloud_cover}
\alias{dark_cloud_cover}
\title{Determine dark cloud cover pattern using a OHLC price series}
\usage{
dark_cloud_cover(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
TRUE if dark cloud cover pattern detected
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
