% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.trend.R
\name{down.trend}
\alias{down.trend}
\title{Determine down trend based on moving average using a OHLC price series}
\usage{
down.trend(x, delta = 0.01, S = 5, L = 20)
}
\arguments{
\item{x}{OHLC prices.}

\item{S}{number of short period for short-run moving average}

\item{L}{number of short period for short-run SMA}
}
\value{
length of upper shadow
}
\description{
Determine down trend based on moving average using a OHLC price series
}
