% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon.doji.R
\name{dragon.doji}
\alias{dragon.doji}
\title{Determine Dragon Doji Pattern using a OHLC price series}
\usage{
dragon.doji(x, delta = 0.1)
}
\arguments{
\item{x}{OHLC prices.}

\item{delta}{Sensivity}
}
\value{
TRUE if Dragon Doji pattern detected
}
\description{
Determine Dragon Doji Pattern using a OHLC price series
}
