% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragonfly.doji.R
\name{dragonfly.doji}
\alias{dragonfly.doji}
\title{Determine Dragon Doji Pattern using a OHLC price series}
\usage{
dragonfly.doji(x, delta = 0.1)
}
\arguments{
\item{x}{OHLC prices.}

\item{delta}{Sensivity}
}
\value{
TRUE if Dragon Doji pattern detected
}
\description{
Determine Dragon Doji Pattern using a OHLC price series
}
