% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_doji.R
\name{dragonfly_doji}
\alias{dragonfly_doji}
\title{Determine Dragon Doji Pattern using a OHLC price series}
\usage{
dragonfly_doji(ohlc, delta = 0.1)
}
\arguments{
\item{delta}{Sensivity}

\item{x}{OHLC prices.}
}
\value{
TRUE if Dragon Doji pattern detected
}
\description{
Determine Dragon Doji Pattern using a OHLC price series
}
