% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evening.star.R
\name{evening.star}
\alias{evening.star}
\title{Determine everning star pattern using a OHLC price series}
\usage{
evening.star(x, n = 20, L.delta = 1, S.delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{L.delta}{sensitivity for long candle}

\item{S.delta}{sensitivity for short candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine everning star pattern using a OHLC price series
}
