% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_star.R
\name{evening_star}
\alias{evening_star}
\title{Determine evening star pattern using a OHLC price series}
\usage{
evening_star(ohlc, n = 20, delta = list(long_candle = 1, short_candle = 1))
}
\arguments{
\item{ohlc}{A \link{data.frame} containing Open, High, Low, Close prices}

\item{n}{number of period of trend}

\item{delta}{A named \link{list} of sensitivity parameters}
}
\value{
TRUE if kicking up pattern detected
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# script: Star-pattern example
# date: 2024-02-12
# author: Serkan Korkmaz, serkor1@duck.com
# objective: A trivial Example
# script start;

# gc
rm(list = ls()); invisible(gc());

# 1) using the pattern function
head(
  candle_pattern(
    ohlc = MSFT,
    pattern = list(
      morning_star(),
      evening_star()
    )
  )
)

# 2) using the function
# itself for each pattern

# 2.1) morning start
head(
  morning_star(
    ohlc = MSFT
  )
)

# 2.2) evening star
head(
  evening_star(
    ohlc = MSFT
  )
)

# script end;
}
\seealso{
Other star: 
\code{\link{morning_star}()}
}
\author{
Chiu-Yu Ko
}
\concept{star}
