% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falling.three.R
\name{falling.three}
\alias{falling.three}
\title{Determine falling three pattern using a OHLC price series}
\usage{
falling.three(x, n = 20, delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{delta}{sensitivity for long candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine falling three pattern using a OHLC price series
}
