% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_three.R
\name{falling_three}
\alias{falling_three}
\title{Determine falling three pattern using a OHLC price series}
\usage{
falling_three(ohlc, n = 20, delta = 1)
}
\arguments{
\item{ohlc}{OHLC prices.}

\item{n}{description}

\item{delta}{sensitivity}
}
\value{
A \link{logical}-vector of same length
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other three-method-patterns: 
\code{\link{rising_three}()}

Other 5-day patterns: 
\code{\link{rising_three}()}
}
\author{
Chiu-Yu Ko
}
\concept{5-day patterns}
\concept{three-method-patterns}
