% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.down.R
\name{gap.down}
\alias{gap.down}
\title{Determine gap down using a OHLC price series}
\usage{
gap.down(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candel detected
}
\description{
Determine gap down using a OHLC price series
}
