% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap.up.R
\name{gap.up}
\alias{gap.up}
\title{Determine gap up using a OHLC price series}
\usage{
gap.up(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candel detected
}
\description{
Determine gap up using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
