% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_gap.R
\name{gap_up}
\alias{gap_up}
\title{Determine gap up using a OHLC price series}
\usage{
gap_up(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
A \link{logical}-vector of same \link{length} with identified gaps.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# script: Gap-pattern example
# date: 2024-02-12
# author: Serkan Korkmaz, serkor1@duck.com
# objective: A trivial Example
# script start;

# gc
rm(list = ls()); invisible(gc());

# 1) using the pattern function
head(
  candle_pattern(
    ohlc = MSFT,
    pattern = list(
      gap_up(),
      gap_down()
    )
  )
)

# 2) using the function
# itself for each pattern

# 2.1) gap down
head(
  gap_down(
    ohlc = MSFT
  )
)

# 2.2) gap up
head(
  gap_up(
    ohlc = MSFT
  )
)

# script end;
}
\seealso{
Other gap: 
\code{\link{gap_down}()}
}
\author{
Chiu-Yu Ko
}
\concept{gap}
