% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravestone.doji.R
\name{gravestone.doji}
\alias{gravestone.doji}
\title{Determine Gravestoen Doji Pattern using a OHLC price series}
\usage{
gravestone.doji(x, delta = 0.1)
}
\arguments{
\item{x}{OHLC prices.}

\item{delta}{Sensivity}
}
\value{
TRUE if Doji pattern detected
}
\description{
Determine Gravestoen Doji Pattern using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
