% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverted.harmer.R
\name{inverted.hammer}
\alias{inverted.hammer}
\title{Determine inverted harmer pattern using a OHLC price series}
\usage{
inverted.hammer(x, US.delta = 0.7, LS.delta = 0.1)
}
\arguments{
\item{x}{OHLC prices.}

\item{US.delta}{sensitivity parameter for upper shadow}

\item{LS.delta}{sensitivity parameter for lower shadow}
}
\value{
TRUE if inverted hammer pattern detected
}
\description{
Determine inverted harmer pattern using a OHLC price series
}
