% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kick.down.R
\name{kick.down}
\alias{kick.down}
\title{Determine kicking down pattern using a OHLC price series}
\usage{
kick.down(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if kicking down pattern detected
}
\description{
Determine kicking down pattern using a OHLC price series
}
