% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kick.up.R
\name{kick.up}
\alias{kick.up}
\title{Determine kicking up pattern using a OHLC price series}
\usage{
kick.up(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine kicking up pattern using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
