% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_kick.R
\name{kick_up}
\alias{kick_up}
\title{Determine kicking up pattern using a OHLC price series}
\usage{
kick_up(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
TRUE if kicking up pattern detected
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
