% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.candle.R
\name{long.candle}
\alias{long.candle}
\title{Determine long candle using a OHLC price series}
\usage{
long.candle(x, n = 20, delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{period}

\item{delta}{sensitivity parameter}
}
\value{
TRUE if long candle detected
}
\description{
Determine long candle using a OHLC price series
}
