% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morning.star.R
\name{morning.star}
\alias{morning.star}
\title{Determine morning star pattern using a OHLC price series}
\usage{
morning.star(x, n = 20, L.delta = 1, S.delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{L.delta}{sensitivity for long candle}

\item{S.delta}{sensitivity for short candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine morning star pattern using a OHLC price series
}
