% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFT.R
\docType{data}
\name{MSFT}
\alias{MSFT}
\title{MSFT}
\format{
An object of class \code{data.frame} with 4306 rows and 6 columns.
}
\usage{
MSFT
}
\description{
A daily OHLC-V xts object between 2007-01-03 and 2024-02-19
}
\details{
\describe{
\item{Open}{Opening price}
\item{High}{Highest price}
\item{Low}{Lowest price}
\item{Close}{Closing price}
\item{Volume}{Volume}
}
}
\concept{data}
\keyword{datasets}
