% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piercing.line.R
\name{piercing.line}
\alias{piercing.line}
\title{Determine piercing line pattern using a OHLC price series}
\usage{
piercing.line(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if hammer pattern detected
}
\description{
Determine piercing line pattern using a OHLC price series
}
