% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_piercing_line.R
\name{piercing_line}
\alias{piercing_line}
\title{Determine piercing line pattern using a OHLC price series}
\usage{
piercing_line(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
TRUE if hammer pattern detected
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
