% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rising.three.R
\name{rising.three}
\alias{rising.three}
\title{Determine rising three pattern using a OHLC price series}
\usage{
rising.three(x, n = 20, delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{delta}{sensitivity for long candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine rising three pattern using a OHLC price series
}
