% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three.black.crows.R
\name{three.black.crows}
\alias{three.black.crows}
\title{Determine three black crows pattern using a OHLC price series}
\usage{
three.black.crows(x, n = 20, delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{delta}{sensitivity for long candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine three black crows pattern using a OHLC price series
}
