% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three.white.soldiers.R
\name{three.white.soldiers}
\alias{three.white.soldiers}
\title{Determine three white soldiers pattern using a OHLC price series}
\usage{
three.white.soldiers(x, n = 20, delta = 1)
}
\arguments{
\item{x}{OHLC prices.}

\item{n}{number of period of trend}

\item{delta}{sensitivity for long candle}
}
\value{
TRUE if kicking up pattern detected
}
\description{
Determine three white soldiers pattern using a OHLC price series
}
