% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_trend.R
\name{trend_down}
\alias{trend_down}
\title{Determine down trend based on moving average using a OHLC price series}
\usage{
trend_down(ohlc, delta = 0.01, S = 5, L = 20)
}
\arguments{
\item{ohlc}{OHLC prices.}

\item{delta}{sensitivity}

\item{S}{number of short period for short-run moving average}

\item{L}{number of short period for short-run SMA}
}
\value{
length of upper shadow
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other Trend patterns: 
\code{\link{trend_up}()}
}
\concept{Trend patterns}
