% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpperShadowLength.R
\name{UpperShadowLength}
\alias{UpperShadowLength}
\title{Determine upper shadow length using a OHLC price series}
\usage{
UpperShadowLength(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
length of upper shadow
}
\description{
Determine upper shadow length using a OHLC price series
}
