% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candle.R
\name{bearish_candle}
\alias{bearish_candle}
\title{Determine bearish candle using a OHLC price series}
\usage{
bearish_candle(ohlc)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candel detected
}
\description{
Determine bearish candle using a OHLC price series
}
