% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engulfing_pattern.R
\name{bearish_engulfing}
\alias{bearish_engulfing}
\title{Determine bearish engulfing pattern using a OHLC price series}
\usage{
bearish_engulfing(ohlc)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bearish engulfing pattern detected
}
\description{
Determine bearish engulfing pattern using a OHLC price series
}
