% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_harami.R
\name{bearish_harami}
\alias{bearish_harami}
\title{Determine bearish harami pattern using a OHLC price series}
\usage{
bearish_harami(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
A \link{logical}-vector of same length
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other harami-patterns: 
\code{\link{bullish_harami}()}

Other 2-day patterns: 
\code{\link{bearish_engulfing}()},
\code{\link{bullish_engulfing}()},
\code{\link{bullish_harami}()}
}
\author{
Chiu-Yu Ko
}
\concept{2-day patterns}
\concept{harami-patterns}
