% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_candle.R
\name{bullish_candle}
\alias{bullish_candle}
\title{Determine bullish candle using a OHLC price series}
\usage{
bullish_candle(ohlc)
}
\arguments{
\item{ohlc}{OHLC prices.}
}
\value{
A \link{logical}-vector of same length of identifed bearish candles
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Chiu-Yu Ko
}
