% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearish.candle.R
\name{bullish_candle}
\alias{bullish_candle}
\title{Determine bullish candle using a OHLC price series}
\usage{
bullish_candle(ohlc, colname = "bullish_candle")
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if bullish candle detected
}
\description{
Determine bullish candle using a OHLC price series
}
