% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engulfing_pattern.R
\name{bullish_engulfing}
\alias{bullish_engulfing}
\title{Determine bullish engulfing pattern using a OHLC price series}
\usage{
bullish_engulfing(ohlc)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if hammer pattern detected
}
\description{
Determine bullish engulfing pattern using a OHLC price series
}
