% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candle_pattern.R
\name{candle_pattern}
\alias{candle_pattern}
\title{Find Candlestick Patterns}
\usage{
candle_pattern(pattern, ohlc, ...)
}
\arguments{
\item{pattern}{A candlestick pattern function}

\item{ohlc}{A xts object with Open, High, Low, Close and Volume columns}

\item{...}{Additional parameters passed to each pattern function}
}
\value{
An xts-object with OHLC-V and the pattern indicator
}
\description{
Find Candlestick Patterns
}
