% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doji.R
\name{doji}
\alias{doji}
\title{Determine Doji Pattern in a OHLC price series}
\usage{
doji(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
TRUE if Doji pattern detected
}
\description{
Determine Doji Pattern in a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
