% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_doji.R
\name{doji}
\alias{doji}
\title{Determine Doji Pattern using a OHLC price series}
\usage{
doji(ohlc, delta = 0.1)
}
\arguments{
\item{ohlc}{OHLC prices.}

\item{delta}{Sensivity}
}
\value{
A \link{logical}-vector of same length
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other doji-patterns: 
\code{\link{dragonfly_doji}()},
\code{\link{gravestone_doji}()}

Other 1-day patterns: 
\code{\link{dragonfly_doji}()},
\code{\link{gravestone_doji}()}
}
\author{
Chiu-Yu Ko
}
\concept{1-day patterns}
\concept{doji-patterns}
