% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_doji.R
\name{doji}
\alias{doji}
\title{Determine Doji Pattern using a OHLC price series}
\usage{
doji(ohlc, delta = 0.1)
}
\arguments{
\item{delta}{sensitivity parameter}

\item{x}{OHLC prices.}
}
\value{
TRUE if Doji pattern detected
}
\description{
Determine Doji Pattern using a OHLC price series
}
