% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_doji.R
\name{gravestone_doji}
\alias{gravestone_doji}
\title{Determine Gravestoen Doji Pattern using a OHLC price series}
\usage{
gravestone_doji(ohlc, delta = 0.1)
}
\arguments{
\item{delta}{Sensivity}

\item{x}{OHLC prices.}
}
\value{
TRUE if Doji pattern detected
}
\description{
Determine Gravestoen Doji Pattern using a OHLC price series
}
