% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer.R
\name{hammer}
\alias{hammer}
\title{Determine harmer pattern using a OHLC price series}
\usage{
hammer(x, US.delta = 0.1, LS.delta = 0.7)
}
\arguments{
\item{x}{OHLC prices.}

\item{US.delta}{sensitivity parameter for upper shadow}

\item{LS.delta}{sensitivity parameter for lower shadow}
}
\value{
TRUE if hammer pattern detected
}
\description{
Determine harmer pattern using a OHLC price series
}
\examples{
getSymbols('AAPL',adjust=TRUE)
doji(AAPL)
}
