% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_hammer.R
\name{hammer}
\alias{hammer}
\title{Identify Hammer patterns in OHLC-series}
\usage{
hammer(ohlc, delta = list(upper_shadow = 0.7, lower_shadow = 0.1))
}
\arguments{
\item{ohlc}{A data.frame containing Open, High, Low, Close prices}

\item{delta}{A named \link{list} of sensitivity parameters
\itemize{
\item upper_shadow
\item lower_shadow
}}
}
\value{
Returns a data.frame-object with identified patterns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other hammer: 
\code{\link{inverted_hammer}()}
}
\author{
Chiu-Yu Ko
}
\concept{hammer}
