% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowerShadowLength.R
\name{LowerShadowLength}
\alias{LowerShadowLength}
\title{Determine lower shadow length using a OHLC price series}
\usage{
LowerShadowLength(x)
}
\arguments{
\item{x}{OHLC prices.}
}
\value{
length of lower shadow;
}
\description{
Determine lower shadow length using a OHLC price series
}
\author{
Ko Chiu Yu, \email{kochiuyu@gmail.com}
}
